+++ NEWS.md
@@ -3,6 +3,7 @@
 * Added coercions from Vector/Mixture/ProductDistribution to Mixture/ProductDistribution via `as.MixtureDistribution` and `as.ProductDistribution`
 * Bugfix in `MixtureDistribution` parameter set
 * Uneven vectors of `WeightedDiscrete` distributions are now truncated at the minimum length in the vector for `pdf, cdf, quantile`
+* `cubature::cubintegrate` now used in `$genExp` method of `CoreStatistics` decorator
 
 # distr6 1.4.3